+++ crates/nu-command/src/date/mod.rs
@@ -1,5 +1,4 @@
 mod date_;
-mod format;
 mod humanize;
 mod list_timezone;
 mod now;
@@ -10,12 +9,10 @@ mod to_timezone;
 mod utils;
 
 pub use date_::Date;
-pub(crate) use format::generate_strftime_list;
-pub use format::SubCommand as DateFormat;
 pub use humanize::SubCommand as DateHumanize;
 pub use list_timezone::SubCommand as DateListTimezones;
 pub use now::SubCommand as DateNow;
 pub use to_record::SubCommand as DateToRecord;
 pub use to_table::SubCommand as DateToTable;
 pub use to_timezone::SubCommand as DateToTimezone;
-pub(crate) use utils::parse_date_from_string;
+pub(crate) use utils::{generate_strftime_list, parse_date_from_string};
