+++ crates/nu-command/src/default_context.rs
@@ -197,7 +197,8 @@ pub fn add_shell_command_context(mut engine_state: EngineState) -> EngineState {
             StrSubstring,
             StrTrim,
             StrTitleCase,
-            StrUpcase
+            StrUpcase,
+            FormatDate
         };
 
         // FileSystem
@@ -233,7 +234,6 @@ pub fn add_shell_command_context(mut engine_state: EngineState) -> EngineState {
         // Date
         bind_command! {
             Date,
-            DateFormat,
             DateHumanize,
             DateListTimezones,
             DateNow,
@@ -300,6 +300,7 @@ pub fn add_shell_command_context(mut engine_state: EngineState) -> EngineState {
         // Env
         bind_command! {
             LetEnvDeprecated,
+            DateFormat,
             ExportEnv,
             LoadEnv,
             SourceEnv,
